requirements.txt
opencv-python
matplotlib
tensorboardX
torchvision
# git+https://github.com/philferriere/cocoapi.git#subdirectory=PythonAPI